backend/src/database/migrations/versions/002_create_tasks_table.py
"""Create tasks table

Revision ID: 002
Revises: 001
Create Date: 2026-01-08

"""
from alembic import op
import sqlalchemy as sa
from sqlalchemy.dialects import postgresql

# revision identifiers, used by Alembic.
revision = '002'
down_revision = '001'
branch_labels = None
depends_on = None


def upgrade() -> None:
    """Create tasks table with foreign key to users."""
    op.create_table(
        'tasks',
        sa.Column('id', postgresql.UUID(as_uuid=True), primary_key=True, server_default=sa.text('gen_random_uuid()')),
        sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=False),
        sa.Column('title', sa.String(length=500), nullable=False),
        sa.Column('description', sa.Text(), nullable=True),
        sa.Column('completed', sa.Boolean(), nullable=False, server_default=sa.text('false')),
        sa.Column('created_at', sa.DateTime(timezone=True), nullable=False, server_default=sa.text('CURRENT_TIMESTAMP')),
        sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False, server_default=sa.text('CURRENT_TIMESTAMP')),
        sa.PrimaryKeyConstraint('id'),
        sa.ForeignKeyConstraint(['user_id'], ['users.id'], name='fk_tasks_user_id', ondelete='CASCADE')
    )

    # Add check constraint to ensure title is not empty
    op.create_check_constraint(
        'check_title_not_empty',
        'tasks',
        sa.text("LENGTH(TRIM(title)) > 0")
    )


def downgrade() -> None:
    """Drop tasks table."""
    op.drop_constraint('check_title_not_empty', 'tasks', type_='check')
    op.drop_table('tasks')
