backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
slowapi==0.1.9
bleach==6.1.0
